requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
gunicorn==23.0.0

torch==2.8.0
opencv-python==4.12.0
numpy==2.2.6
Pillow==11.3.0

diffusers==0.31.0
transformers==4.46.0
accelerate==0.34.2
safetensors==0.6.2
huggingface_hub==0.34.4
